+++ R/Jackknife.R
@@ -37,7 +37,7 @@ Jackknife <- function(tree, dataset, resampleFreq = 2 / 3,
   morphyObj <- InitializeData(dataset)
   on.exit(morphyObj <- CleanUpData(morphyObj))
   
-  startWeights <- MorphyWeights(morphyObj)["exact", ]
+  startWeights <- .MorphyWeight(morphyObj)
   eachChar <- seq_along(startWeights)
   deindexedChars <- rep.int(eachChar, startWeights)
   charsToKeep <- ceiling(resampleFreq * length(deindexedChars))
@@ -61,16 +61,8 @@ Jackknife <- function(tree, dataset, resampleFreq = 2 / 3,
     } #nocov end
     resampling <- tabulate(sample(deindexedChars, charsToKeep, replace = FALSE),
                            nbins = length(startWeights))
-    errors <- vapply(eachChar, function (i) 
-      mpl_set_charac_weight(i, resampling[i], morphyObj), integer(1))
-    if (any(errors)) { #nocov start
-      stop ("Error resampling morphy object: ", 
-            mpl_translate_error(unique(errors[errors < 0L])))
-    }
-    if (mpl_apply_tipdata(morphyObj) -> error) {
-      stop("Error applying tip data: ", mpl_translate_error(error))
-    } #nocov end
-    res <- EdgeListSearch(edgeList[1:2], morphyObj, EdgeSwapper = EdgeSwapper,
+    resampledObj <- .SetMorphyWeight(morphyObj, resampling)
+    res <- EdgeListSearch(edgeList[1:2], resampledObj, EdgeSwapper = EdgeSwapper,
                           maxIter = searchIter, maxHits = searchHits,
                           verbosity = verbosity - 1L, ...)
     res[1:2]